+++ examples/typescript-react-example/src/ImportHookDefault.gen.tsx
@@ -9,28 +9,28 @@ import {default as defaultNotChecked} from './hookExample';
 // In case of type error, check the type of 'make' in 'ImportHookDefault.re' and './hookExample'.
 export const makeTypeChecked: React.ComponentType<{
   readonly person: person; 
-  readonly children: JSX.Element; 
+  readonly children: React.ReactNode; 
   readonly renderMe: ImportHooks_renderMe<string>
 }> = makeNotChecked;
 
 // Export 'make' early to allow circular import from the '.bs.js' file.
 export const make: unknown = makeTypeChecked as React.ComponentType<{
   readonly person: person; 
-  readonly children: JSX.Element; 
+  readonly children: React.ReactNode; 
   readonly renderMe: ImportHooks_renderMe<string>
 }>;
 
 // In case of type error, check the type of 'default' in 'ImportHookDefault.re' and './hookExample'.
 export const defaultTypeChecked: React.ComponentType<{
   readonly person: person; 
-  readonly children: JSX.Element; 
+  readonly children: React.ReactNode; 
   readonly renderMe: ImportHooks_renderMe<string>
 }> = defaultNotChecked;
 
 // Export '$$default' early to allow circular import from the '.bs.js' file.
 export const $$default: unknown = defaultTypeChecked as React.ComponentType<{
   readonly person: person; 
-  readonly children: JSX.Element; 
+  readonly children: React.ReactNode; 
   readonly renderMe: ImportHooks_renderMe<string>
 }>;